remove unused type, and move UserItem to a separate component
