+++ package.json
@@ -7,7 +7,7 @@
     "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint ./components ./App.js",
+    "lint": "eslint ./components ./App.js ./types.js",
     "flow": "flow",
     "flow-stop": "flow stop"
   },
